# managed-identity-wallets-app/src/main/resources/static/openapi_v2.yaml
openapi: 3.0.0
info:
  title: Verifiable Credentials API
  version: 2.0.0
  description: An API for managing Verifiable Credentials and DID Documents.
servers:
  - url: /api/v2
paths:
...
  /verifiable-credentials/{verifiableCredentialId}:
    get:
      operationId: userGetVerifiableCredentialById
      tags:
        - User
      summary: Get Verifiable Credential by ID
      description: Retrieves a verifiable credential from the wallet by ID.
      parameters:
        - $ref: '#/components/parameters/Path_Verifiable_Credential_Id'
      responses:
        '200':
          description: Successful response with the requested verifiable credential.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiableCredentialResponsePayload'
        '404':
          description: Verifiable credential not found.

    delete:
      operationId: userDeleteVerifiableCredentialById
      tags:
        - User
      summary: Delete Verifiable Credential
      description: Removes Verifiable Credential from the wallet by ID.
      parameters:
        - $ref: '#/components/parameters/Path_Verifiable_Credential_Id'
      responses:
        '204':
          description: Verifiable Credential deleted successfully.
        '404':
          description: Verifiable Credential not found.
...
  /signed-verifiable-presentations/jwt:
    post:
      operationId: userIssuedVerifiablePresentationJwt
      tags:
        - User
      summary: Issue Verifiable Presentation as JWT
      description: Issues a new Verifiable Presentation as JWT.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueVerifiablePresentationJwtRequestPayload'
      responses:
        '200':
          description: Verifiable Presentation issued successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueVerifiablePresentationJwtResponsePayload'
        '400':
          description: Bad request.
...
  /admin/verifiable-credentials/{verifiableCredentialId}:
    get:
      operationId: adminGetVerifiableCredentialById
      tags:
        - Administrator
      summary: Get Verifiable Credentials by Id
      description: Retrieves a list of verifiable credentials for a user by user ID.
      parameters:
        - $ref: '#/components/parameters/Path_Verifiable_Credential_Id'
      responses:
        '200':
          description: Successful response with a verifiable credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiableCredentialResponsePayload'
        '404':
          description: Verifiable credential not found.


    delete:
      operationId: adminDeleteVerifiableCredentialById
      tags:
        - Administrator
      summary: Delete Verifiable Credential
      description: Deletes a verifiable credential
      parameters:
        - $ref: '#/components/parameters/Path_Verifiable_Credential_Id'
      responses:
        '204':
          description: Verifiable credential deleted successfully.
        '404':
          description: Verifiable credential not found.
components:
  parameters:
...
    Path_Verifiable_Credential_Id:
      name: verifiableCredentialId
      in: path
      required: true
      description: The ID of the verifiable credential.
      schema:
        type: string
...
  schemas:
...
    VerifiableCredentialResponsePayload:
      description: key value map
      type: object
      additionalProperties: true
...
    IssueVerifiablePresentationJwtRequestPayload:
      description: Verifiable Credential Request Payload
      type: object
      required: [ audience, verifiableCredentials ]
      properties:
        audience:
          description: The audience for the JWT
          type: string
        verifiableCredentials:
          description: Verifiable Credentials
          type: array
          items:
            description: key value map
            type: string
            additionalProperties: true
    IssueVerifiablePresentationJwtResponsePayload:
      description: Verifiable Presentation as JWT
      type: object
      properties:
        verifiable-presentation:
          description: Verifiable Presentation as JWT
          type: string
...
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Grants read access
            write: Grants write access
            admin: Grants access to admin operations